ARTSurfaceViewShadowNode: Avoid not drawing the Surface View
Calls updateExtraData(), which registers the surface texture listener may arrive too late,
thus the SurfaceTexture may already have been created which will case onSurfaceTextureAvailable()
to never be called, ultimatilly causing the surface to not be drawed.

As a fix this patch checks if the SurfaceTexture is ready to use and if so draw the surface right away.